GOVERNMENT CHATBOT/requirements.txt
flask
flask-cors
transformers
torch
sentence-transformers
faiss-cpu
scikit-learn